Lagt til forlengetSaksbehandling og refactoring
